requirements.txt
django>=1.6.2,<1.7
sae-python-dev>=1.3.2